Reorganizing files based on functionality to better understand the separation of responsibilities.
