+++ app/src/main/java/com/amohnacs/bitcointv/details/ui/DetailsActivity.java
@@ -12,23 +12,27 @@
  * the License.
  */
 
-package com.amohnacs.bitcointv;
+package com.amohnacs.bitcointv.details.ui;
 
+import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
 
-/**
- * Loads {@link PlaybackVideoFragment}.
+import com.amohnacs.bitcointv.R;
+
+/*
+ * Details activity class that loads LeanbackDetailsFragment class
  */
-public class PlaybackActivity extends FragmentActivity {
+public class DetailsActivity extends Activity {
+    public static final String SHARED_ELEMENT_NAME = "hero";
+    public static final String MOVIE = "Movie";
 
+    /**
+     * Called when the activity is first created.
+     */
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (savedInstanceState == null) {
-            getSupportFragmentManager().beginTransaction()
-                    .replace(android.R.id.content, new PlaybackVideoFragment())
-                    .commit();
-        }
+        setContentView(R.layout.activity_details);
     }
+
 }
